Update patch set 1
Patch Set 1:

I'm not a fan of the freq called list as I use that third page in Dialer as my Favorites/All Contacts list instead - so I do see some value in the functionality. 

That said, if it is broken and not respecting it's state, then it should be fixed. 

I actually brought this up the other day as something we should default to 'on' and remove the option altogether, because in my head the freq called list == call log. 

But Koush (like you) balked at that idea as he uses that functionality. S

Patch-set: 1