Update patch set 2
Patch Set 2:

Exported contacts are stored in the contactsprovider like any other contact. At load time while we know its an exported SIM contact, not sure if it makes sense to not show this contact all together. Though it does make sense to not show the exported contact in the multiselect delete list.

Furthermore, even if we were to prevent it from showing up in the delete list, this fix is still needed to prevent a FC condition. This code path should be protected.

Patch-set: 2