Update patch set 3
Patch Set 3:

@Michal
This patch is trying to address that issue. Before, 'local phone account' would only show up if you had 2 or more accounts. With this patch, you will always see 'local phone' as an option. Are you saying that this patch still doesn't fix the issue for you ?

Patch-set: 3