Update patch set 1
Patch Set 1: Code-Review+1

As this has been included in former version of CM as well as in the Marshmallow version of OMNIROM without any problems I vote for the inclusion of this diff.

Assuming that the external account type supports groups even if it does not seems to be better than the opposite.

Reference:

https://forums.bitfire.at/topic/129/category-group-field-for-contacts-is-not-synchronized/7

Patch-set: 1