Update patch set 1
Patch Set 1: Code-Review+1

Right now, I have to use the Xposed module EnableContactsGroups to be able to sync the groups with my caldav server using DavDroid. I couldn't see any impact because of this.
As the xposed module does the same as the patch, I'd like to see this merged.

Patch-set: 1