Update patch set 3
Patch Set 3: Code-Review-1

> In this case, we can set CheckableQuickContact to set this to true
 > instead of setting the default to true and only false on the
 > imageview. I like that solution better. Sound good?

Actually, no. The CheckedImageView also is used for photos in a lot of cases, which is the reason why I converted it to use CheckedFlipDrawable in the first place.

My preferred solution is http://review.cyanogenmod.org/#/c/142096/ as it constrains the non-scaling path to drawable resources.

Patch-set: 3
Label: Code-Review=-1